// src/enums/GameType.ts
export enum GameType {
  Blackjack = 'Blackjack',
  Roulette = 'Roulette',
  Craps = 'Craps',
  PaiGow = 'PaiGow',
  Baccarat = 'Baccarat',
  ThreeCardPoker = 'ThreeCardPoker',
  TexasHoldem = 'TexasHoldem',
  UltimateTexasHoldem = 'UltimateTexasHoldem',
  MississippiStud = 'MississippiStud',
  SpanishTwentyOne = 'SpanishTwentyOne',
  Other = 'Other'
}